Package names are mentioned along with repo name now
Therefore, error reporting has changed a bit. Now, the package name
is mentioned as ${P}::repo_name and the script may fail to recognize
those packages which failed. An example of this is opencascade which
requires a specific amount of space in /var/tmp/portage, and aborts
if that space is not present.